Adapted unit tests
